backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "decision-simulation-engine-backend"
version = "0.1.0"
description = "Backend for the Decision Simulation Engine"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2,<2.0.0",
  "beautifulsoup4>=4.12.3,<5.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.27.2,<1.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "psycopg[binary]>=3.2.1,<4.0.0",
  "SQLAlchemy>=2.0.36,<2.1.0",
  "uvicorn[standard]>=0.32.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "black>=24.10.0,<25.0.0",
  "pytest>=8.3.3,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "ruff>=0.7.2,<1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["app/tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.black]
line-length = 100
